requirements.txt
fastapi==0.115.9
uvicorn==0.34.0
pandas
joblib
scikit-learn
xgboost
openai
pydantic
python-dotenv
requests